jump search

+++ 0x1E-search_algorithms/100-jump.c
@@ -15,7 +15,7 @@ int jump_search(int *array, size_t size, int value)
 	if (array == NULL || size == 0)
 		return (-1);
 
-	step = sqrt(size);
+	step = (int)sqrt((double)size);
 	prev = 0;
 
 	while (array[step] < value && step < size)
